requirements.txt
flask
requests
yfinance
python-dotenv
sqlalchemy
edgartools
fredapi
anthropic
